Update Paciente.java
Se realizan cambios en los tipos de datos de los getter de numero y cedula, aparecian como "double", y el tipo de dato es "long"

+++ Hospital/src/co/ean/hospital/Paciente.java
@@ -24,11 +24,11 @@ public class Paciente {
         return nombre;
     }
 
-    public double getCedula() {
+    public long getCedula() {
         return cedula;
     }
 
-    public double getNumero() {
+    public long getNumero() {
         return numero;
     }
